# PriorityQueue: Apply Comparer<T>.Default optimization