Make direct calls to `Comparer<T>.Default.Compare` when no custom `IComparer<T>` instance is specified, taking advantage of the optimizations in #48160.

## Performance

Using benchmarks from https://github.com/dotnet/performance/pull/1665

### main

|         Method |   Size |            Mean |         Error |        StdDev |          Median |             Min |             Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |------- |----------------:|--------------:|--------------:|----------------:|----------------:|----------------:|------:|------:|------:|----------:|
|       HeapSort |     10 |        275.5 ns |       4.72 ns |       3.94 ns |        274.3 ns |        272.3 ns |        284.7 ns |     - |     - |     - |         - |
| K_Min_Elements |     10 |        188.2 ns |       1.43 ns |       1.34 ns |        188.1 ns |        186.4 ns |        190.5 ns |     - |     - |     - |         - |
|       HeapSort |    100 |      5,831.1 ns |      23.68 ns |      20.99 ns |      5,827.2 ns |      5,790.5 ns |      5,873.7 ns |     - |     - |     - |         - |
| K_Min_Elements |    100 |        835.3 ns |       5.56 ns |       4.64 ns |        835.9 ns |        825.7 ns |        842.4 ns |     - |     - |     - |         - |
|       HeapSort |   1000 |    104,618.2 ns |   1,297.35 ns |   1,213.54 ns |    104,295.8 ns |    103,134.7 ns |    107,482.9 ns |     - |     - |     - |         - |
| K_Min_Elements |   1000 |      5,109.5 ns |      84.29 ns |      74.72 ns |      5,082.6 ns |      5,035.9 ns |      5,285.3 ns |     - |     - |     - |         - |
|       HeapSort |  10000 |  1,382,286.0 ns |  17,339.01 ns |  14,478.85 ns |  1,378,520.1 ns |  1,364,851.3 ns |  1,418,781.0 ns |     - |     - |     - |       1 B |
| K_Min_Elements |  10000 |     43,600.4 ns |     655.10 ns |     547.04 ns |     43,470.5 ns |     42,982.0 ns |     44,965.3 ns |     - |     - |     - |         - |
|       HeapSort | 100000 | 17,338,945.6 ns | 127,125.13 ns | 106,155.24 ns | 17,342,100.0 ns | 17,143,023.1 ns | 17,486,884.6 ns |     - |     - |     - |      11 B |
| K_Min_Elements | 100000 |    461,078.6 ns |  18,155.17 ns |  20,907.53 ns |    458,911.1 ns |    428,690.8 ns |    492,106.5 ns |     - |     - |     - |         - |

### PR branch

|         Method |   Size |            Mean |         Error |        StdDev |          Median |             Min |             Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |------- |----------------:|--------------:|--------------:|----------------:|----------------:|----------------:|------:|------:|------:|----------:|
|       HeapSort |     10 |        208.9 ns |       6.34 ns |       6.79 ns |        208.6 ns |        191.8 ns |        217.7 ns |     - |     - |     - |         - |
| K_Min_Elements |     10 |        147.3 ns |       5.80 ns |       6.68 ns |        144.6 ns |        136.6 ns |        160.8 ns |     - |     - |     - |         - |
|       HeapSort |    100 |      3,553.9 ns |     120.23 ns |     133.63 ns |      3,583.6 ns |      3,261.7 ns |      3,751.3 ns |     - |     - |     - |         - |
| K_Min_Elements |    100 |        610.5 ns |      18.93 ns |      20.25 ns |        615.2 ns |        561.9 ns |        647.7 ns |     - |     - |     - |         - |
|       HeapSort |   1000 |     82,091.4 ns |   3,752.99 ns |   4,321.94 ns |     82,769.6 ns |     74,755.7 ns |     91,248.6 ns |     - |     - |     - |         - |
| K_Min_Elements |   1000 |      4,042.3 ns |     179.86 ns |     207.13 ns |      4,028.6 ns |      3,687.8 ns |      4,411.0 ns |     - |     - |     - |         - |
|       HeapSort |  10000 |  1,117,819.2 ns |  38,606.25 ns |  42,910.75 ns |  1,108,901.0 ns |  1,037,257.2 ns |  1,176,602.9 ns |     - |     - |     - |       1 B |
| K_Min_Elements |  10000 |     34,393.6 ns |   1,911.55 ns |   2,124.68 ns |     33,527.5 ns |     31,757.1 ns |     38,687.1 ns |     - |     - |     - |         - |
|       HeapSort | 100000 | 14,617,127.5 ns | 510,218.11 ns | 587,567.94 ns | 14,697,175.0 ns | 13,688,916.7 ns | 15,793,872.2 ns |     - |     - |     - |       8 B |
| K_Min_Elements | 100000 |    337,853.6 ns |  12,592.12 ns |  14,501.10 ns |    331,689.4 ns |    316,677.5 ns |    365,371.4 ns |     - |     - |     - |         - |