@eiriktsarpalis 2021-02-17
> Do we have sufficient tests now that we need to cover value types vs reference types with and without a custom comparer and where that comparer is and is not the default?

I think we could do better. I'll follow up with another PR.